Moved email verification to settings sidebar; Added email to sidebar; Turning persistent login ON
