+++ thebikekollective/lib/components/home_body.dart
@@ -46,12 +46,6 @@ class _HomeBodyState extends State<HomeBody> {
                     SizedBox(height: spacerHeight),
                     placeholderBoxes(
                         'Add Bike Placeholder', addBikeWidth, addBikeHeight),
-                    SizedBox(height: spacerHeight),
-                    emailVerificationButton(
-                        context,
-                        'Email Verification (this will move)',
-                        searchBarWidth,
-                        searchBarHeight),
                     SizedBox(height: spacerHeight)
                   ]),
             )));
@@ -76,75 +70,5 @@ class _HomeBodyState extends State<HomeBody> {
     mapHeight = screenHeight * 0.45;
     addBikeWidth = screenWidth * 0.50;
     addBikeHeight = screenHeight * 0.05;
-  }
-
-  Widget emailVerificationButton(BuildContext context, String text,
-      double buttonWidth, double buttonHeight) {
-    return ElevatedButton(
-        onPressed: () async {
-          SharedPreferences preferences = await SharedPreferences.getInstance();
-          String? username = preferences.getString('username');
-          if (username == 'no username') {
-            ScaffoldMessenger.of(context).showSnackBar(testUserSnackBar());
-          } else {
-            var snapshot = await FirebaseFirestore.instance
-                .collection('users')
-                .where('username', isEqualTo: username)
-                .get();
-            snapshot.docs.forEach((result) {
-              if (result.data()['verified'] == true) {
-                print('EMAIL ALREADY VERIFIED');
-                ScaffoldMessenger.of(context)
-                    .showSnackBar(alreadyVerifiedSnackBar());
-              } else {
-                print('EMAIL IS NOT VERIFIED');
-                Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) => EmailVerificationScreen()));
-              }
-            });
-          }
-        },
-        child: emailVerificationButtonText(text),
-        style: ElevatedButton.styleFrom(
-            primary: Color(s_periwinkleBlue),
-            fixedSize: Size(buttonWidth, buttonHeight)));
-  }
-
-  Widget emailVerificationButtonText(String text) {
-    return FormattedText(
-      text: text,
-      size: s_fontSizeSmall,
-      color: Colors.white,
-      font: s_font_AmaticSC,
-      weight: FontWeight.bold,
-    );
-  }
-
-  SnackBar alreadyVerifiedSnackBar() {
-    return SnackBar(
-        backgroundColor: Color(s_jungleGreen),
-        content: FormattedText(
-          text: 'Email is already verified!',
-          size: s_fontSizeSmall,
-          color: Colors.white,
-          font: s_font_BonaNova,
-          weight: FontWeight.bold,
-          align: TextAlign.center,
-        ));
-  }
-
-  SnackBar testUserSnackBar() {
-    return SnackBar(
-        backgroundColor: Color(s_declineRed),
-        content: FormattedText(
-          text: 'You are using the test user bypass!',
-          size: s_fontSizeSmall,
-          color: Colors.white,
-          font: s_font_BonaNova,
-          weight: FontWeight.bold,
-          align: TextAlign.center,
-        ));
   }
 }
