+++ thebikekollective/lib/main.dart
@@ -4,8 +4,7 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'app.dart';
 
-
-Future<void> main() async{
+Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   SystemChrome.setPreferredOrientations([
     DeviceOrientation.landscapeLeft,
@@ -15,9 +14,9 @@ Future<void> main() async{
   await Firebase.initializeApp();
   SharedPreferences preferences = await SharedPreferences.getInstance();
 
-  // Manual sign out (Turn this off to enable persistent login)
-  preferences.setBool('loggedIn', false);
-  preferences.setString('username', 'no username');
+  // Manual sign out (Turn this on to DISABLE persistent login)
+  // preferences.setBool('loggedIn', false);
+  // preferences.setString('username', 'no username');
 
   print(preferences.getBool('loggedIn'));
   print(preferences.getString('username'));
